move math foreign_items into their own file
